Added two things: 1) I noticed that variables cannot be named with Capital letters and was getting error. So I changed the lex file(as suggested by Geoff).
2) While creating an AssignStmt it should check whether the Expr
is Proc type or not. If it is proc type then it stores the current
NameTable in the Proc class in a private NameTable (as Geoff was discussing today.) . This way the current environment is stored.
In the apply of the Proc along with the NT the restored NameTable is copied into the NNT for the evaluation of the StmtList.
x:=0;
y:=0;
a := proc(x) return:= proc(y) return:= x+y end end;
b:=a(2);
c:=a(3);
d:=b(10);
e:=c(10)
Here when the function b is called with a parameter it should apply 10 to y and x should be 2. Because while function b was created (as a result of the return value from funccall a(2) it should retain the value 2 for x.).